requirements.txt
django==5.1.1
djangorestframework
markdown
django-filter
djangorestframework-simplejwt